requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
imutils>=0.5.4
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.8
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
websockets>=12.0  # WebSocket support
python-dateutil>=2.8.2  # Date handling for analytics
sqlalchemy>=2.0.23  # Database ORM
pydantic[email]>=2.5.1  # Data validation with email support
aiosmtplib>=2.0.2  # Async SMTP for email notifications
httpx>=0.25.1  # Async HTTP for webhooks